add commands package
